gcf-function/requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
black==21.4b2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
fonttools==4.29.1
future==0.18.2
fvcore==0.1.5.post20220119
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-functions==1.5.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
iopath==0.1.9
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
portalocker==2.3.2
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
yacs==0.1.8
zipp==3.7.0
